# services/api/openapi.yaml
...
openapi: 3.0.3
info:
  title: ft_transcendence API
  version: 1.1.0
  description: >
    API for the ft_transcendence project.  
    Provides authentication (JWT + rotating refresh cookies), health monitoring, and game-related endpoints.
servers:
  - url: http://localhost:5000
    description: Local development server
...
paths:
...
  /api/auth/refresh:
    post:
      summary: Refresh access token (rotate refresh)
      description: >
        Uses the httpOnly refresh cookie to validate and **rotate** the refresh token, then returns a new access token.  
        No Authorization header required. The endpoint sets a new rotated refresh cookie.
      tags: [Authentication]
      operationId: refresh
      responses:
        "200":
          description: New access token issued
          headers:
            Set-Cookie:
              description: >
                Rotated httpOnly refresh token cookie (cookie name via `COOKIE_NAME_RT`).
              schema:
                type: string
          content:
            application/json:
              schema:
                type: object
                properties:
                  succes:
                    type: boolean
                    example: true
                  token:
                    type: string
                    description: New JWT access token
                  exp:
                    type: string
                    example: "15m"
        "401":
          description: Unauthorized (missing/invalid/expired/rotated elsewhere)
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT